airflow/requirements.txt
aiofiles
aiohttp
aiosignal
alembic
anyio
apache-airflow
apache-airflow-providers-common-sql
apache-airflow-providers-ftp
apache-airflow-providers-google
apache-airflow-providers-http
apache-airflow-providers-imap
apache-airflow-providers-sqlite
apispec
argcomplete
asgiref
async-timeout
attrs
Babel
backoff
blinker
cachelib
cachetools
cattrs
certifi
cffi
chardet
charset-normalizer
click
clickclick
colorama
colorlog
commonmark
ConfigUpdater
connexion
cron-descriptor
croniter
cryptography
db-dtypes
Deprecated
dill
dnspython
docutils
email-validator
exceptiongroup
Flask
Flask-AppBuilder
Flask-Babel
Flask-Caching
Flask-JWT-Extended
Flask-Login
Flask-Session
Flask-SQLAlchemy
Flask-WTF
frozenlist
future
gcloud-aio-auth
gcloud-aio-bigquery
gcloud-aio-storage
google-ads
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-aiplatform
google-cloud-appengine-logging
google-cloud-audit-log
google-cloud-automl
google-cloud-bigquery
google-cloud-bigquery-datatransfer
google-cloud-bigquery-storage
google-cloud-bigtable
google-cloud-build
google-cloud-compute
google-cloud-container
google-cloud-core
google-cloud-datacatalog
google-cloud-dataflow-client
google-cloud-dataform
google-cloud-dataplex
google-cloud-dataproc
google-cloud-dataproc-metastore
google-cloud-dlp
google-cloud-kms
google-cloud-language
google-cloud-logging
google-cloud-memcache
google-cloud-monitoring
google-cloud-orchestration-airflow
google-cloud-os-login
google-cloud-pubsub
google-cloud-redis
google-cloud-resource-manager
google-cloud-secret-manager
google-cloud-spanner
google-cloud-speech
google-cloud-storage
google-cloud-tasks
google-cloud-texttospeech
google-cloud-translate
google-cloud-videointelligence
google-cloud-vision
google-cloud-workflows
google-crc32c
google-resumable-media
googleapis-common-protos
graphviz
greenlet
grpc-google-iam-v1
grpcio
grpcio-gcp
grpcio-status
gunicorn
h11
httpcore
httplib2
httpx
idna
inflection
itsdangerous
Jinja2
json-merge-patch
jsonschema
lazy-object-proxy
linkify-it-py
lockfile
looker-sdk
Mako
Markdown
markdown-it-py
MarkupSafe
marshmallow
marshmallow-enum
marshmallow-oneofschema
marshmallow-sqlalchemy
mdit-py-plugins
mdurl
multidict
numpy
oauthlib
packaging
pandas
pandas-gbq
pathspec
pendulum
pip-chill
pluggy
prison
proto-plus
protobuf
psutil
pyarrow
pyasn1
pyasn1-modules
pycparser
pydata-google-auth
Pygments
PyJWT
pyOpenSSL
pyparsing
pyrsistent
python-daemon
python-dateutil
python-nvd3
python-slugify
pytz
pytzdata
PyYAML
requests
requests-oauthlib
requests-toolbelt
rfc3986
rich
rsa
setproctitle
six
sniffio
SQLAlchemy
sqlalchemy-bigquery
SQLAlchemy-JSONField
SQLAlchemy-Utils
sqlparse
swagger-ui-bundle
tabulate
tenacity
termcolor
text-unidecode
typing_extensions
uc-micro-py
unicodecsv
uritemplate
urllib3
Werkzeug
wrapt
WTForms
yarl
